requirements.txt
amqp==1.0.13
anyjson==0.3.3
asgiref==3.2.10
astroid==2.4.2
autopep8==1.4.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
bootstrap4==0.1.0
celery==2.5.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.41.0
django-bootstrap4==3.0.1
django-bulma==0.8.2
django-celery-beat==1.5.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-taggit==1.3.0
django-timezone-field==4.1.1
et-xmlfile==1.0.1
Flask==2.0.2
fonttools==4.29.1
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.3
joblib==0.16.0
jsonify==0.5
kiwisolver==1.2.0
kombu==2.5.16
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
MouseInfo==0.1.3
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.19.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==21.3
pandas==1.0.5
pep8==1.7.1
pi==0.1.2
Pillow==7.2.0
prompt-toolkit==3.0.8
psycopg2==2.8.5
PyAutoGUI==0.9.52
pycodestyle==2.5.0
PyGetWindow==0.0.9
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.1.0
PyTweening==1.0.3
pytz==2020.1
pywhatkit==3.9
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
soupsieve==2.1
sqlparse==0.3.1
tablib==3.0.0
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.1.0
wikipedia==1.4.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0